refactor(app): simplify process qs
- moved out handle_message_edit
- moved out QsStreamProcessor into a dedicated module
- added resync required helper
- factored out chat and group loading
- removed 4-tuple destructuring and renamed
  ApplicationMessagesHandlerResult to HandledMessages
- factor out resolve_group_profile_part
- removed never read changed_chats field
- replaced ProcessedQsMessages with a simpler QsMessageOutcome
